refactor: memberService Api 인터페이스와 구현체로 분리
