+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/create/CreateArticleService.java
@@ -4,5 +4,5 @@ import foot.footprint.domain.article.dto.CreateArticleRequest;
 
 public interface CreateArticleService {
 
-    public Long create(CreateArticleRequest request, Long memberId);
+    Long create(CreateArticleRequest request, Long memberId);
 }
