+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/delete/DeleteArticleService.java
@@ -2,5 +2,5 @@ package foot.footprint.domain.article.application.delete;
 
 public interface DeleteArticleService {
 
-    public void delete(Long articleId, Long memberId);
+    void delete(Long articleId, Long memberId);
 }
