+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/edit/EditArticleService.java
@@ -2,5 +2,5 @@ package foot.footprint.domain.article.application.edit;
 
 public interface EditArticleService {
 
-    public void edit(Long articleId, Long memberId, String newContent);
+    void edit(Long articleId, Long memberId, String newContent);
 }
