+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/findArticles/FindArticlesService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface FindArticlesService {
 
-    public List<ArticleMapResponse> findArticles(Long memberId, Long groupId,
+    List<ArticleMapResponse> findArticles(Long memberId, Long groupId,
         LocationRange locationRange);
 }
