+++ backend/footprint/src/main/java/foot/footprint/domain/articleLike/application/ChangeArticleLikeService.java
@@ -4,5 +4,5 @@ import foot.footprint.domain.articleLike.dto.ArticleLikeDto;
 
 public interface ChangeArticleLikeService {
 
-    public void changeArticleLike(ArticleLikeDto articleLikeDto);
+    void changeArticleLike(ArticleLikeDto articleLikeDto);
 }
