+++ backend/footprint/src/main/java/foot/footprint/domain/comment/application/delete/DeleteCommentService.java
@@ -4,5 +4,5 @@ import foot.footprint.domain.article.exception.NotMatchMemberException;
 
 public interface DeleteCommentService {
 
-    public void delete(Long commentId, Long memberId);
+    void delete(Long commentId, Long memberId);
 }
