+++ backend/footprint/src/main/java/foot/footprint/domain/comment/application/edit/EditCommentService.java
@@ -2,5 +2,5 @@ package foot.footprint.domain.comment.application.edit;
 
 public interface EditCommentService {
 
-    public void edit(Long commentId, Long memberId, String newContent);
+    void edit(Long commentId, Long memberId, String newContent);
 }
